Move mock helpers to separate file
